Move monitoring radios into their own module
This is part of splitting up the large parsl.monitoring.monitoring
module (933 lines). Radios need to live somewhere they can be
accessed by a soon-to-be moved remote monitoring wrapper, without
importing the full monitoring module.
